+++ NebulaWorld/MonoBehaviours/Remote/RemotePlayerMovement.cs
@@ -30,6 +30,7 @@ namespace NebulaWorld.MonoBehaviours.Remote
         private GameObject playerDot;
         private GameObject playerName;
         public String Username { get; set; }
+        public ushort PlayerID { get; set; }
 
 #if DEBUG
         private GameObject positionDebugger;
@@ -53,13 +54,10 @@ namespace NebulaWorld.MonoBehaviours.Remote
             if (origPlayerDot != null && uiSailIndicator_targetText != null)
             {
                 playerDot = Instantiate(origPlayerDot, origPlayerDot.transform.parent, false);
-
                 playerName = Instantiate(origPlayerDot, origPlayerDot.transform.parent, false);
-                //playerName.transform.SetParent(playerDot.transform, false);
 
                 Destroy(playerName.GetComponent<MeshFilter>());
 
-                //MeshRenderer meshRenderer = playerName.AddComponent<MeshRenderer>();
                 MeshRenderer meshRenderer = playerName.GetComponent<MeshRenderer>();
                 playerName.AddComponent<TextMesh>();
 
@@ -104,30 +102,33 @@ namespace NebulaWorld.MonoBehaviours.Remote
                 return;
             }
 
+            // update navigation indicator if requested
+            if(GameMain.mainPlayer.navigation.indicatorAstroId > 100000)
+            {
+                UpdateNavigationGizmo();
+            }
+
             // update player dot on minimap if on same planet
             if(playerDot != null && playerName != null && localPlanetId == GameMain.mainPlayer.planetId)
             {
                 // compute spherical distance from us to player to hide his name if he is on the other side of the planet
                 double distance = PlayerNavigation.SphericalDistance(GameCamera.main.transform.position, rootTransform.position, GameMain.localPlanet.realRadius, false);
-
-                if (distance >= GameMain.localPlanet.realRadius)
-                {
-                    playerDot.SetActive(false);
-                    playerName.SetActive(false);
-                }
-                else
-                {
-                    playerDot.SetActive(true);
-                    playerName.SetActive(true);
-
-                    playerDot.transform.localPosition = rootTransform.position * (float)(0.5 / (double)GameMain.localPlanet.realRadius);
-                    playerDot.transform.localScale = 0.02f * Vector3.one;
-
-                    playerName.transform.localPosition = playerDot.transform.localPosition;
-                    playerName.transform.rotation = UIRoot.instance.uiGame.planetGlobe.minimapControl.cam.transform.rotation;
-                }
-
                 TextMesh textMesh = playerName.GetComponent<TextMesh>();
+
+                playerDot.SetActive(true);
+                playerName.SetActive(true);
+
+                if (textMesh != null)
+                {
+                    textMesh.color = new Color(textMesh.color.r, textMesh.color.g, textMesh.color.b, (distance >= GameMain.localPlanet.realRadius) ? 0.2f : 1f);
+                }
+
+                playerDot.transform.localPosition = rootTransform.position * (float)(0.5 / (double)GameMain.localPlanet.realRadius);
+                playerDot.transform.localScale = 0.02f * Vector3.one;
+
+                playerName.transform.localPosition = playerDot.transform.localPosition;
+                playerName.transform.rotation = UIRoot.instance.uiGame.planetGlobe.minimapControl.cam.transform.rotation;
+
                 if (textMesh != null && textMesh.text != Username)
                 {
                     TextMesh uiSailIndicator_targetText = UIRoot.instance.uiGame.sailIndicator.targetText;
@@ -171,6 +172,57 @@ namespace NebulaWorld.MonoBehaviours.Remote
                     // This will skip interpolation and will snap to the most recent position.
                     MoveInterpolated(snapshotBuffer[i], snapshotBuffer[i + 1], 1);
                 }
+            }
+        }
+
+        private void UpdateNavigationGizmo()
+        {
+            if(PlayerID == GameMain.mainPlayer.navigation.indicatorAstroId - 100000)
+            {
+                PlayerControlGizmo gizmo = GameMain.mainPlayer.gizmo;
+                UIStarmap starmap = UIRoot.instance.uiGame.starmap;
+
+                if (gizmo.naviIndicatorGizmo == null)
+                {
+                    if (gizmo.player.navigation.indicatorAstroId > 0)
+                    {
+                        gizmo.naviIndicatorGizmo = LineGizmo.Create(1, gizmo.player.position, rootTransform.position);
+                    }
+                    gizmo.naviIndicatorGizmo.autoRefresh = true;
+                    gizmo.naviIndicatorGizmo.multiplier = 1.5f;
+                    gizmo.naviIndicatorGizmo.alphaMultiplier = 0.3f;
+                    gizmo.naviIndicatorGizmo.width = 0.6f;
+                    gizmo.naviIndicatorGizmo.color = Configs.builtin.gizmoColors[4];
+                    gizmo.naviIndicatorGizmo.spherical = false;
+                    gizmo.naviIndicatorGizmo.Open();
+                }
+                if (gizmo.naviIndicatorGizmoStarmap == null)
+                {
+                    if (gizmo.player.navigation.indicatorAstroId > 0)
+                    {
+                        gizmo.naviIndicatorGizmoStarmap = LineGizmo.Create(1, gizmo.player.position, rootTransform.position);
+                    }
+                    gizmo.naviIndicatorGizmoStarmap.autoRefresh = true;
+                    gizmo.naviIndicatorGizmoStarmap.multiplier = 1.5f;
+                    gizmo.naviIndicatorGizmoStarmap.alphaMultiplier = 0.3f;
+                    gizmo.naviIndicatorGizmoStarmap.width = 0.01f;
+                    gizmo.naviIndicatorGizmoStarmap.color = Configs.builtin.gizmoColors[4];
+                    gizmo.naviIndicatorGizmoStarmap.spherical = false;
+                    gizmo.naviIndicatorGizmoStarmap.Open();
+                }
+
+                gizmo.naviIndicatorGizmo.startPoint = gizmo.player.position;
+                if (gizmo.player.navigation.indicatorAstroId > 0)
+                {
+                    gizmo.naviIndicatorGizmo.endPoint = rootTransform.position;
+                }
+                gizmo.naviIndicatorGizmoStarmap.startPoint = (gizmo.player.uPosition - starmap.viewTargetUPos) * 0.00025;
+                if (gizmo.player.navigation.indicatorAstroId > 0)
+                {
+                    gizmo.naviIndicatorGizmoStarmap.endPoint = (absolutePosition - starmap.viewTargetUPos) * 0.00025;
+                }
+
+                gizmo.naviIndicatorGizmoStarmap.gameObject.layer = 20;
             }
         }